renamde submodules from plural to singular term
