+++ cognite/neat/graph/extractor/_base.py
