+++ cognite/neat/graph/extractor/_graph_capturing_sheet.py
