+++ cognite/neat/graph/transformation/query_generator/sparql.py
