+++ tests/tests_unit/conftest.py
@@ -6,9 +6,9 @@ from pathlib import Path
 import pytest
 from rdflib import RDF, Literal, Namespace
 
-from cognite.neat.graph import extractors, loaders
+from cognite.neat.graph import extractor, loader
 from cognite.neat.graph.stores import MemoryStore
-from cognite.neat.graph.transformations.transformer import domain2app_knowledge_graph
+from cognite.neat.graph.transformation.transformer import domain2app_knowledge_graph
 from cognite.neat.rules import importer
 from cognite.neat.rules.exporter._rules2triples import get_instances_as_triples
 from cognite.neat.rules.models.rules import Rules
@@ -146,7 +146,7 @@ def mock_knowledge_graph(transformation_rules) -> MemoryStore:
         "Terminal": 2,
     }
 
-    mock_triples = extractors.MockGraphGenerator(transformation_rules, class_count).extract()
+    mock_triples = extractor.MockGraphGenerator(transformation_rules, class_count).extract()
     mock_graph.add_triples(mock_triples, batch_size=20000)
 
     return mock_graph
@@ -154,12 +154,12 @@ def mock_knowledge_graph(transformation_rules) -> MemoryStore:
 
 @pytest.fixture(scope="function")
 def mock_rdf_assets(mock_knowledge_graph, transformation_rules):
-    return loaders.rdf2assets(mock_knowledge_graph, transformation_rules, data_set_id=123456)
+    return loader.rdf2assets(mock_knowledge_graph, transformation_rules, data_set_id=123456)
 
 
 @pytest.fixture(scope="function")
 def mock_cdf_assets(mock_knowledge_graph, transformation_rules):
-    return loaders.rdf2assets(mock_knowledge_graph, transformation_rules, data_set_id=123456)
+    return loader.rdf2assets(mock_knowledge_graph, transformation_rules, data_set_id=123456)
 
 
 @pytest.fixture(scope="function")
