+++ tests/tests_unit/graph/extractors/test_sheet2graph.py
@@ -1,6 +1,6 @@
 from rdflib import XSD, Literal
 
-from cognite.neat.graph import extractors
+from cognite.neat.graph import extractor
 from cognite.neat.graph.stores import MemoryStore
 from cognite.neat.utils.utils import remove_namespace
 from tests import config
@@ -10,7 +10,7 @@ def test_sheet2graph(simple_rules):
     graph_store = MemoryStore(prefixes=simple_rules.prefixes, namespace=simple_rules.metadata.namespace)
     graph_store.init_graph(base_prefix=simple_rules.metadata.prefix)
 
-    triples = extractors.GraphCapturingSheet(simple_rules, config.GRAPH_CAPTURING_SHEET, use_source_ids=True).extract()
+    triples = extractor.GraphCapturingSheet(simple_rules, config.GRAPH_CAPTURING_SHEET, use_source_ids=True).extract()
 
     graph_store.add_triples(triples)
