+++ tests/tests_unit/graph/transformations/test_transformer.py
@@ -2,7 +2,7 @@ import pandas as pd
 from cognite.client.testing import monkeypatch_cognite_client
 from rdflib import Graph, Namespace
 
-from cognite.neat.graph.transformations.transformer import domain2app_knowledge_graph
+from cognite.neat.graph.transformation.transformer import domain2app_knowledge_graph
 from cognite.neat.rules.models.rdfpath import TransformationRuleType
 from cognite.neat.rules.models.rules import Rules
