+++ tests/tests_unit/rules/test_rdfpath.py
@@ -6,7 +6,7 @@ from pydantic import ValidationError
 
 from cognite.neat.constants import PREFIXES
 from cognite.neat.graph.stores import MemoryStore
-from cognite.neat.graph.transformations.query_generator import build_sparql_query
+from cognite.neat.graph.transformation.query_generator import build_sparql_query
 from cognite.neat.rules.models.rdfpath import (
     AllProperties,
     AllReferences,
